// src/types.d.ts
type RoleTypes = 'admin' | 'guest' | 'student' | 'professor' | 'customer' | 'evaluator' | 'moderator'

type CourseTypes = 'Ciência da Computação' | 'Engenharia da Computação' | 'Engenharia Mecânica' | 'Engenharia de Produção' | 'Sistemas de Informação'

type Pagination = { page: number; per_page?: number }